requirements.txt
Django
django-lockdown
Pillow
djangorestframework
sorl-thumbnail
ipython